unit testing, refactor
