+++ rapid7_insightidr/komand_rapid7_insightidr/actions/advanced_query_on_log/action.py
@@ -4,9 +4,7 @@ from komand.exceptions import PluginException
 # Custom imports below
 import time
 import json
-from dateutil.parser import parse, ParserError
-from komand_rapid7_insightidr.util.relative_time_codes import relative_time_to_milliseconds
-
+from komand_rapid7_insightidr.util.parse_dates import parse_dates
 
 class AdvancedQueryOnLog(komand.Action):
 
@@ -27,7 +25,7 @@ class AdvancedQueryOnLog(komand.Action):
         time_to_string = params.get(Input.TIME_TO)
 
         # Time To is optional, if not specified, time to is set to now
-        time_from, time_to = self.parse_dates(time_from_string, time_to_string, relative_time_from)
+        time_from, time_to = parse_dates(time_from_string, time_to_string, relative_time_from)
 
         log_id = self.get_log_id(log_name)
 
@@ -46,29 +44,7 @@ class AdvancedQueryOnLog(komand.Action):
         self.logger.info(f"Sending results to orchestrator.")
         return {Output.RESULTS: log_entries}
 
-    def parse_dates(self, time_from_string: str, time_to_string: str, relative_time_from: str) -> (int, int):
-        """
-        Parse incoming dates and return them as millisecond epoch time
-
-        @param time_from_string: str
-        @param time_to_string: str (optional, if it's a falsey value, time to will be set to Now)
-        @return: (int, int)
-        """
-        rel_time_milli = relative_time_to_milliseconds.get(relative_time_from)
-
-        # Parse times to epoch milliseconds
-        try:
-            time_from = int(parse(time_from_string).timestamp()) * 1000
-            if time_to_string:
-                time_to = (int(parse(time_to_string).timestamp()) * 1000) - rel_time_milli
-            else:
-                # Now in millisecond epoch time
-                time_to = (int(time.time()) * 1000) - rel_time_milli
-        except ParserError as e:
-            raise PluginException(cause="Could not parse given date.",
-                                  assistance="The date given was in an unrecognizable format.",
-                                  data=e)
-        return time_from, time_to
+        # Parse Dates
 
     def get_results_from_callback(self, callback_url: str, timeout: int) -> [object]:
         """
